Render-to-texture option
Trying to figure out how to best render to texture.  May just use the
texture coordinates to write to a separate UAV that can be retained
between passes.  Have a dummy render target for the first pass, then not
clear the UAV before I use it for the next part.